Add experimental 'Forge patch' mixins set to fix remap bug
Forge adds mod jars to the classpath immediately before loading them.
However, FMLDeobfuscatingRemapper relies on having all classes available
on the classpath to properly handle references to obfuscated fields in superclasses (i.e. 'super maps')

This creates an issue when a mod references classes from a companion
mod. When the mod containing the reference is remapped,
FMLDeobfuscatingRemapper will attempt to generate the 'super map' for
the class located in the other mod. Since the other mod is not yet on
the classpath, 'super map' generation will fail, and this failure will
be cached. When the other mod is later classloaded, this cached lookup
failure will prevent remapping from correctly occuring.